Injecting system calls into alarm manager
Injecting and mocking dependencies to make the service unit testable.
Shuffled around initialization to make that possible.

Also added a native method to query timerfd_gettime. This is helpful in
debugging because it will tell for certain whether the kernel timer was
armed or not and whether it matches the expectation of higher level code
in the service.

Also If an alarm delivery is stuck, detect and log a wtf. This will help us
catch relevant information sooner and also indicate how frequently this
happens. Otherwise this is something that can easily go unnoticed and is
usually hard to debug once the alarms are running as expected again.

Bug: 111454659
Bug: 78560047
Bug: 78560047

Change-Id: Ifa1c09646f688cf5c41abe17d98dbc19978eac70
Change-Id: If08ebf69b0bde7e8f504d81b98e3f41ecf7f2dc2
Change-Id: Ife881f2c9649b8dfd87ba05a425483375361da64